core: extract capture content helpers into src/core/capture-content.ts
Pure move of slug defaulting (inbox/YYYY-MM-DD-sha8 + type routing), the CV10
binary NUL guard, CV9 hash normalization, title derivation, the Life Chronicle
event block, and the BUG-1 frontmatter merge — shared with the upcoming
capture MCP op. capture.ts statically imports operations.ts, so operations-
layer code can never import capture.ts; the core module breaks that cycle.
capture.ts re-exports the public trio (detectBinaryNullByte, normalizeForHash,
mergeCaptureFrontmatter) and keeps its __testing seam, stdin reader, and the
FK-hint rewriter (CLI/thin-client-only concerns).
